+++ src/backend/métier/entities/entité-juridique/ÉtablissementTerritorialRattaché.ts
@@ -5,4 +5,6 @@ export type ÉtablissementTerritorialRattaché = Readonly<{
   numéroFiness: string;
   raisonSocialeCourte: string;
   libelléCatégorieÉtablissement: string;
+  libelleCourtCategorieEtablissement: string;
+  categorieEtablissementCode: string;
 }>;