+++ src/frontend/ui/entité-juridique/liste-des-établissements/ÉtablissementTerritorialRattachéViewModel.tsx
@@ -20,6 +20,14 @@ export class EtablissementTerritorialRattacheViewModel {
     return this.etablissementTerritorialRattache.libelléCatégorieÉtablissement;
   }
 
+  public get libelleCourtCategorieEtablissement(): string {
+    return this.etablissementTerritorialRattache.libelleCourtCategorieEtablissement;
+  }
+
+    public get categorieEtablissementCode(): string {
+    return this.etablissementTerritorialRattache.categorieEtablissementCode;
+  }
+
   public get identifiant(): ReactElement {
     return (
       <>